requirements.txt
anyio==3.6.2
pyjwt>=2.3.0
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt== 4.0.1
beautifulsoup4==4.10.0
beniget==0.4.1
black==23.3.0
blinker==1.4
Brotli== 1.0.9
certifi==2021.5.30
cffi== 1.15.1
chardet==4.0.0
click==8.0.1
colorama==0.4.4
configobj==5.0.6
constantly==    15.1.0
cryptography==  40.0.2
cycler== 0.11.0
dbus-python==   1.2.18
decorator==4.4.2
distro== 1.7.0
et-xmlfile==1.1.0
fastapi==0.66.0
fonttools==4.29.1
fs==   2.4.12
gast== 0.5.2
greenlet==1.1.0
h11==  0.12.0
html5lib==1.1
httplib2==0.20.2
httptools==0.5.0
hyperlink==21.0.0
idna== 3.4
importlib-metadata==4.6.4
incremental==   21.3.0
ipython==7.31.1
jedi== 0.18.0
jeepney==0.7.1
Jinja2== 3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
more-itertools==8.10.0
mypy-extensions==  0.4.3
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.7
packaging==23.1
parso==0.8.1
passlib==1.7.4
pathspec==0.11.1
pexpect==4.8.0
pickleshare==   0.7.5
Pillow== 9.0.1
pip==  22.0.2
platformdirs==  3.5.1
ply==  3.11
prompt-toolkit==3.0.28
ptyprocess==    0.7.0
pyasn1== 0.4.8
pyasn1-modules==0.2.1
pycparser==2.20
pydantic==1.8.2
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==    2.0.2
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==    0.18.1
pyserial==3.5
python-dateutil==  2.8.1
python-dotenv== 1.0.0
python-magic==  0.4.24
pythran==0.10.0
pytz== 2022.1
PyYAML== 6.0
regex==2021.7.6
requests==2.25.1
scipy==1.8.0
SecretStorage== 3.3.1
service-identity== 18.1.0
setuptools==    59.6.0
six==  1.16.0
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==    1.4.20
ssh-import-id== 5.11
starlette==0.14.2
sympy==1.9
toml== 0.10.2
tomli==2.0.1
traitlets==5.1.1
ubuntu-advantage-tools==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==  14.0.0
urllib3==1.26.5
uvicorn==0.14.0
uvloop== 0.17.0
wadllib==1.3.6
watchfiles== 0.19.0
wcwidth==0.2.5
webencodings==0.5.1
websockets== 11.0
wheel==0.37.1
zipp== 1.0.0
zope.interface==5.4.0